// Pandem2-Prototype-FE/src/app/core/models/modelling-data.model.ts
export class ModellingModelParameterValue implements IModellingModelParameterValue {
  value?: number | boolean;
  limits?: {
    min?: number;
    max?: number;
  };
  age?: ModellingModelParameterValueAgeType;
  ageContact?: ModellingModelParameterValueAgeContactType;

  constructor(data = null) {
    this.value = _.get(data, 'value');
    this.limits = _.get(data, 'limits');
    this.age = _.get(data, 'age');
    this.age = _.get(data, 'ageContact');
  }
}

export class ModellingModelParameter implements IModellingModelParameter {
  name: string;
  key: string;
  category: ModellingModelParameterCategory;
  subcategory?: ModellingModelParameterSubcategory;
  description?: string;
  type: ModellingModelParameterType;
  step?: number;
  readonly?: boolean;
  values: ModellingModelParameterValue[];

  constructor(data = null) {
    this.name = _.get(data, 'name');
    this.key = _.get(data, 'key');
    this.category = _.get(data, 'category');
    this.subcategory = _.get(data, 'subcategory');
    this.description = _.get(data, 'description');
    this.type = _.get(data, 'type');
    this.step = _.get(data, 'step');
    this.readonly = _.get(data, 'readonly');
    this.values = _.get(data, 'values');
  }
}

export class ModellingModelDescriptionSection implements IModellingModelDescriptionSection {
  tab_title: string;
  items: Array<IModellingModelDescriptionSimpleText
  | IModellingModelDescriptionImage
  | IModellingModelDescriptionTable
  | IModellingModelDescriptionTextArray>;

  constructor(data = null) {
    this.tab_title = _.get(data, 'tab_title');
    this.items = _.get(data, 'items');
  }
}

export class ModellingModel implements ModellingModelDataEntity {
  id: string;
  name: string;
  key: string;
  pathogen: string;
  short_description?: string;
  description?: ModellingModelDescriptionSection[];
  model_structure_image?: string;
  parameters: ModellingModelParameter[];
  parameters_updated_at: string;

  constructor(data = null) {
    this.id = _.get(data, '_id');
    this.name = _.get(data, 'name');
    this.key = _.get(data, 'key');
    this.pathogen = _.get(data, 'pathogen');
    this.short_description = _.get(data, 'short_description');
    this.description = _.get(data, 'description');
    this.model_structure_image = _.get(data, 'model_structure_image');
    this.parameters = _.get(data, 'parameters');
    this.parameters_updated_at = _.get(data, 'parameters_updated_at');
  }
}

export class ModellingScenarioParameterValue implements IModellingScenarioParameterValue {
  value: number | boolean;
  age?: ModellingModelParameterValueAgeType;
  ageContact?: ModellingModelParameterValueAgeContactType;

  constructor(data = null) {
    this.value = _.get(data, 'value');
    this.age = _.get(data, 'age');
    this.ageContact = _.get(data, 'ageContact');
  }
}

export class ModellingScenarioParameter implements IModellingScenarioParameter {
  key: string;
  values: ModellingScenarioParameterValue[];

  constructor(data = null) {
    this.key = _.get(data, 'key');
    this.values = _.get(data, 'values');
  }
}

export class ModellingAlternativeScenario implements IModellingAlternativeScenario {
  id: string;
  name: string;

  constructor(data = null) {
    this.id = _.get(data, 'id');
    this.name = _.get(data, 'name');
  }
}

export class ModellingScenario implements IModellingScenarioDataEntity {
  id: string;
  userId: string;
  modelId: string;
  name: string;
  r0: number;
  date: Date;
  description?: string;
  tags: string[];
  location: string;
  parameters: ModellingScenarioParameter[];
  sections_details?: IModellingSectionDetails[];
  exploration?: ModellingExplorationChartData[];
  is_visible?: boolean;
  alternatives?: ModellingAlternativeScenario[];

  constructor(data = null) {
    this.id = _.get(data, '_id');
    this.userId = _.get(data, 'userId');
    this.modelId = _.get(data, 'modelId');
    this.name = _.get(data, 'name');
    this.r0 = _.get(data, 'r0');
    this.date = _.get(data, 'date');
    this.description = _.get(data, 'description');
    this.tags = _.get(data, 'tags');
    this.location = _.get(data, 'location');
    this.parameters = _.get(data, 'parameters');
    this.sections_details = _.get(data, 'sections_details');
    this.exploration = _.get(data, 'exploration');
    this.is_visible = _.get(data, 'is_visible');
    this.alternatives = _.get(data, 'alternatives');
  }
}
// ...
export class ModellingExplorationChartData implements IModellingExplorationChartData {
  chart_type: 'spline' | 'column' | 'area';
  chart_plot_type: LinearLog;
  view_by: 'scenario' | 'indicator';
  view_style: ModellingViewType;
  values: string[];
  plotlines: string[];

  constructor(data = null) {
    this.chart_type = _.get(data, 'chart_type');
    this.chart_plot_type = _.get(data, 'chart_plot_type');
    this.view_by = _.get(data, 'view_by');
    this.view_style = _.get(data, 'view_style');
    this.values = _.get(data, 'values');
    this.plotlines = _.get(data, 'plotlines');
  }
}

export class ModellingScenarioWithDayResults implements IModellingScenarioWithDayResultsDataEntity {
  id: string;
  userId: string;
  modelId: string;
  name: string;
  r0: number;
  date: Date;
  description?: string;
  tags: string[];
  location: string;
  parameters: ModellingScenarioParameter[];
  sections_details?: IModellingSectionDetails[];
  day_results?: ModellingScenarioDayResult[];
  processed_results?: IModellingScenarioProcessedDayResult;
  exploration?: ModellingExplorationChartData[];
  is_visible?: boolean;
  alternatives?: ModellingAlternativeScenario[];

  constructor(data = null) {
    this.id = _.get(data, 'id');
    this.userId = _.get(data, 'userId');
    this.modelId = _.get(data, 'modelId');
    this.name = _.get(data, 'name');
    this.r0 = _.get(data, 'r0');
    this.date = _.get(data, 'date');
    this.description = _.get(data, 'description');
    this.tags = _.get(data, 'tags');
    this.location = _.get(data, 'location');
    this.parameters = _.get(data, 'parameters');
    this.sections_details = _.get(data, 'sections_details');
    this.day_results = _.get(data, 'day_results');
    this.processed_results = _.get(data, 'processed_results');
    this.exploration = _.get(data, 'exploration');
    this.is_visible = _.get(data, 'is_visible');
    this.alternatives = _.get(data, 'alternatives');
  }
}